GoCrackIT/requirements.txt
Flask
Flask-Cors
requests
werkzeug
python-dotenv
google-generativeai
assemblyai
PyPDF2
python-docx
pandas
Pillow
pytesseract 